+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## 1.1.0
+
+- Added hover instruction documentation
+- Added hover register documentation
+- Added hover binary/decimal/hex/char convertion
+- Added constant recognition and definition (control+click jump) features
+- Added label recognition and definition (control+click jump) features
+
 ## 1.0.0
 
 - Initial release: syntax highlighting for VS Code